Add local dates to elastic search documents
Elastic search stores dates in UTC and Kibana displays dates in the client time zone. While it is acceptable in many cases, trip and flight dates are bounded to airports and thus it would be nice to see dates in airport time zones. As a workaround, additionally to date properties, the string representation of these dates in airport time zones have been added. For example, for trip_date property of type "date" a new property trip_date_local of type string have been added which represents the date in the trip origin airport.