Define trip period as ranges of departure and return dates
Before the trip period in config file was defined by departure dates and trip duration. Now, it requires the range of return dates to be specified explicitly:
{
    "fromAirport": [
        "SYD"
    ],
    "toAirport": [
        "KBP"
    ],
    "minDate": "09.12.2016",
    "maxDate": "12.12.2016",
    "minReturnDate": "10.01.2017",
    "maxReturnDate": "13.01.2017",
    "minDuration": 32
}

Such approach provides more flexibility in defining return dates for the trip. In the same time, it is still possible to specify the min duration of the trip.